Add a second's delay to t7519 for untracked cache
In t7519 there is a test that writes files to disk, and immediately
writes the untracked index. Because of mtime-comparison logic
that uses a 1-second resolution, this means the cached entries
are not trusted/used under some circumstances (see
read-cache.c#is_racy_stat()).

Untracked cache tests in t7063 use a 1-second delay to avoid
this issue. We should do the same here.

This change doesn't actually affect the outcome of the test,
but does enhance its validity, and becomes relevant after
later changes

Signed-off-by: Tao Klerks <tao@klerks.biz>