Write index when populating empty untracked cache
It is expected that an empty/unpopulated untracked
cache structure can be written to the index - by update-
index, or by a "git status" call that sees the untracked cache
should be enabled, but is running with options that make
it non-applicable in that run.

Currently, if that happens, then subsequent "git status"
calls end up populating the untracked cache, but not
writing the index (not saving their work) - so the
performance outcome is almost identical to the cache
being altogether disabled.

This continues until the index gets written with the cache
populated, for some *other* reason.

In this change, we detect the "existing cache is empty
and it looks like we are using it" condition, and queues
an index write when this happens.

This change depends on previous fixes to t7519 for the
"ignore .git changes when invalidating UNTR" test case to
pass - before this fix, the test never actually did anything
as it was not set up correctly.

Signed-off-by: Tao Klerks <tao@klerks.biz>