merge-tree: provide an easy way to access which files have conflicts
Callers of `git merge-tree --real` might want an easy way to determine
which files conflicted.  While they could potentially use the --messages
option and parse the resulting messages written to that file, those
messages are not meant to be machine readable.  Provide a simpler
mechanism of having the user specify --unmerged-list=$FILENAME, and then
write a NUL-separated list of unmerged filenames to the specified file.

Signed-off-by: Elijah Newren <newren@gmail.com>