[GSOC] ref-filter: introducing xstrvfmt_len() and xstrfmt_len()
In the logic of ref-filter, if we already specified v->s_size (the
length of the data), we can avoid using strlen() to calculate the
length of the data in strbuf_addstr() or perl_quote_buf_with_len()...

xstrfmt(), xstrvfmt(), they are functions used to add format strings
to a dynamic allocated buffer. Because their implementation uses strbuf
to store its buffer and its length, so we can easily pass its length
to the caller, furthermore, in ref-filter, we can reduce the use of
strlen().

So introduce xstrvfmt_len() and xstrfmt_len() which can fetch the length
of the buffer through the parameter `len`. This can bring a slight
performance improvement.

Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: Hariom Verma <hariom18599@gmail.com>
Signed-off-by: ZheNing Hu <adlternative@gmail.com>