[GSOC] ref-filter: introducing xstrvfmt_len() and xstrfmt_len()
xstrfmt(), xstrvfmt(), they are functions used to add formatt strings
to a dynamic memory. Because their implementation uses strbuf to store
its buffer and its length, so we can easily pass its length to the caller,
so create xstrvfmt_len() and xstrfmt_len() which can fetch the length
of the buffer through the parameter `len`. This can reduce the cost of
calculating the length of the string later through strlen().

Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: Hariom Verma <hariom18599@gmail.com>
Signed-off-by: ZheNing Hu <adlternative@gmail.com>