[GSOC] ref-filter: reduce unnecessary object_info comparisons
In populate_value() we use an empty struct object_info
compare with oi and oi_deref, which determine if we can
return early, this is actually very inefficient.

So add a need_get_object_info flag in global variables,
which used to indicate whether we need call get_object()
or we can return early.

So we only need to compare only one byte instead of comparing
the entire huge object_info struct.

Mentored-by: Christian Couder <christian.couder@gmail.com>
Mentored-by: Hariom Verma <hariom18599@gmail.com>
Signed-off-by: ZheNing Hu <adlternative@gmail.com>